applied keep to lat and lon files

+++ src/scripts/mod02.pl
@@ -209,8 +209,10 @@ my $row_file  = "$tag\_rows_$swath_cols\_$swath_rows.img";
 
 do_or_die("ll2cr -v $swath_cols $swath_rows $lat_file $lon_file " .
 	  "$gpdfile $col_file $row_file");
-do_or_die("rm -f $lat_file");
-do_or_die("rm -f $lon_file");
+if (!$keep) {
+    do_or_die("rm -f $lat_file");
+    do_or_die("rm -f $lon_file");
+}
 
 open_or_die("GPDFILE", "$ENV{PATHMPP}/$gpdfile");
 my $line = <GPDFILE>;
